Update day4_love_score_calc.py
This program automates a "True Love" quiz from a magazine. The goal was to practice count, lower, and control flow. The first version was not totaling properly. I thought that it was a problem with my algorithm, but it turns out my string concatenation was actually just totaling the True and Love and integers and not scoring correctly. After the second cleaner algorithm didn't work I started testing all my totals and found the issue. A lot of learning happened for such a simple program.