Create day7_who_buys.py
In this file, I was practicing using lists and random to create a small application that randomly selects a name from a list to determine who among a group of friends would buy lunch. it is using the split method to take a user input and create a list, then len to determine a range for a random number generator to pick a position.