Create day9_FizzBuzz
This was really easy, especially after Rock, Paper, Scissors. All I did here was generate a range 1-100, loop through and if n is not divisible by 3 or 5 print it. Then if divisible by 3 AND 5 print FizzBuzz. Next, I checked 3 to print Fizz and 5 to print Buzz. Super fun mental exercise.